+++ src/jdk.compiler/share/classes/com/sun/tools/javac/comp/Lower.java
@@ -109,7 +109,6 @@ public class Lower extends TreeTranslator {
     private final boolean useMatchException;
     private final HashMap<TypePairs, String> typePairToName;
     private int variableIndex = 0;
-    private final boolean dumpStacktraceOnError;
 
     @SuppressWarnings("this-escape")
     protected Lower(Context context) {
@@ -142,7 +141,6 @@ public class Lower extends TreeTranslator {
         useMatchException = Feature.PATTERN_SWITCH.allowedInSource(source) &&
                             (preview.isEnabled() || !preview.isPreview(Feature.PATTERN_SWITCH));
         typePairToName = TypePairs.initialize(syms);
-        dumpStacktraceOnError = options.isSet("dev") || options.isSet(DOE);
     }
 
     /** The currently enclosing class.
@@ -2633,7 +2631,7 @@ public class Lower extends TreeTranslator {
         StringBuilder sb = new StringBuilder();
 
         LowerSignatureGenerator() {
-            super(types, Lower.this.dumpStacktraceOnError);
+            types.super();
         }
 
         @Override
