+++ src/jdk.compiler/share/classes/com/sun/tools/javac/jvm/PoolWriter.java
@@ -100,7 +100,7 @@ public class PoolWriter {
     public PoolWriter(Types types, Names names) {
         this.types = types;
         this.names = names;
-        this.signatureGen = new SharedSignatureGenerator(types);
+        this.signatureGen = new SharedSignatureGenerator();
         this.pool = new WriteablePoolHelper();
     }
 
@@ -278,8 +278,8 @@ public class PoolWriter {
          */
         ByteBuffer sigbuf = new ByteBuffer();
 
-        SharedSignatureGenerator(Types types) {
-            super(types, PoolWriter.this.types.dumpStacktraceOnError);
+        SharedSignatureGenerator() {
+            types.super();
         }
 
         /**
