+++ src/jdk.jshell/share/classes/jdk/jshell/TreeDissector.java
@@ -246,7 +246,7 @@ class TreeDissector {
         StringBuilder sb = new StringBuilder();
 
         TDSignatureGenerator(Types types) {
-            super(types, types.dumpStacktraceOnError);
+            types.super();
         }
 
         @Override
